ml/requirements.txt
# Core Dependencies
neo4j>=5.0.0
pandas>=1.3.0
python-dotenv>=0.19.0

# PyTorch with CUDA (install via):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126


# PyTorch and PyTorch Geometric
torch>=2.5.0
torch-geometric>=2.7.0

# Spark + GraphFrames (graph builder pipeline)
pyspark>=3.5.0
graphframes>=0.6

# Scientific Computing
numpy>=1.21.0
scikit-learn>=1.0.0

# Jupyter and Visualization
jupyter>=1.0.0
matplotlib>=3.4.0
ipykernel>=6.0.0
plotly>=5.0.0

# Development Tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Note: torch-geometric dependencies
# The following may be needed depending on your PyTorch version:
# torch-scatter
# torch-sparse
# torch-cluster
# torch-spline-conv
# 
# Install instructions for PyTorch Geometric dependencies:
# For PyTorch 2.5.x with CUDA 11.8:
#   pip install torch-scatter torch-sparse torch-cluster torch-spline-conv -f https://data.pyg.org/whl/torch-2.5.0+cu118.html
# 
# For other versions, visit: https://pytorch-geometric.readthedocs.io/en/latest/install/installation.html

